+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This update makes some changes to the **planet pack** configs. Please make sure 
 * Fix RSS reentry starting too late
 * Add error/problem handling and list, warning users about incorrect installs or other issues
 * Add warning about enabled stock effects
+* Added additional message to the effect editor, informing to not use the simulation sliders in normal gameplay
 
 ## 1.0.0.0 - 2025-02-05